add multipage apps
